Should fix the search issue (from dirty commit from k/nitrous)
This is the SHA of the dirty commit on kartik/nitrous:
98f4f594937d38cc71d9afe22fbb7682fa845c0b

If there is still a problem, this is the commit that is causing it so
you can dismiss all code changes made in this commit. Easiest hacky way
to do this is to just copy-paste or checkout files from the other
branch and specify the SHA (can google the syntax) if no changes have
been made, else merge across commits/branches.